Refactor meshplot methods to submodule
